fix(env): clean imports for tests, remove dependencies, fix depracated warnings
