+++ libs/renderer/testing/block-defaults/LinkBlock.spec.tsx
@@ -2,7 +2,8 @@ import { expect } from "vitest";
 import "@testing-library/jest-dom";
 
 import { render, screen } from "../utils";
-import { LinkBlock } from "../../components/block-defaults/link-block/LinkBlock";
+
+import { LinkBlock } from "@/components/block-defaults/link-block/LinkBlock";
 
 const blocks = {
     link: {
