+++ libs/renderer/vitest.setup.ts
@@ -1,2 +1,2 @@
 import "@testing-library/jest-dom";
-import 'vitest-canvas-mock';
+import "vitest-canvas-mock";
