+++ packages/client/vite.config.mts
@@ -18,6 +18,7 @@ export default defineConfig({
     resolve: {
         alias: [{ find: '@', replacement: resolve(__dirname, './src') }],
     },
+
     // Uncomment this if you are using workers.
     // worker: {
     //  plugins: [ nxViteTsPaths() ],
@@ -30,8 +31,11 @@ export default defineConfig({
         environment: 'jsdom',
         include: ['**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
         deps: {
-            // Required for vitest-canvas-mock
-            inline: ['vitest-canvas-mock'],
+            optimizer: {
+                web: {
+                    include: ['vitest-canvas-mock'],
+                },
+            },
         },
         reporters: ['default'],
         coverage: {
@@ -42,9 +46,6 @@ export default defineConfig({
             jsdom: {
                 resources: 'usable',
             },
-        },
-        cache: {
-            dir: '../../node_modules/.vitest',
         },
         setupFiles: ['./vitest.setup.ts'],
     },
